Major updates to rad2bt and bt2rad, consolidated location in aslutil with radtot and ttorad set as symlinks.
This fixes the old and new problems and includes Paul's idea for
extending dimensions past 2 and Breno's to allow fr to be scalar for
any shape of rad.  It should be backwards compatible with the older
versions (unless you were depending on the bugs) and is shorter and
simpler than the 2009 version.